Update Example Writer, renamed to Fatal Writer, Add plaintext version
