+++ spec/PhpSpec/Process/Shutdown/UpdateConsoleActionSpec.php
@@ -3,6 +3,7 @@
 namespace spec\PhpSpec\Process\Shutdown;
 
 use PhpSpec\Formatter\CurrentExampleWriter;
+use PhpSpec\Formatter\WriterInterface;
 use PhpSpec\Message\CurrentExample;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
@@ -10,22 +11,12 @@ use Prophecy\Argument;
 class UpdateConsoleActionSpec extends ObjectBehavior
 {
 
-    function let(CurrentExample $currentExample, CurrentExampleWriter $currentExampleWriter)
+    function let(CurrentExample $currentExample, WriterInterface $currentExampleWriter)
     {
         $this->beConstructedWith($currentExample, $currentExampleWriter);
     }
 
-    function it_is_initializable()
-    {
-        $this->shouldHaveType('PhpSpec\Process\Shutdown\UpdateConsoleAction');
-    }
-
-    function it_should_extend_shutdown_action_interface()
-    {
-        $this->shouldHaveType('PhpSpec\Process\Shutdown\ShutdownActionInterface');
-    }
-
-    function it_should_update_the_console(CurrentExample $currentExample, CurrentExampleWriter $currentExampleWriter)
+    function it_should_update_the_console(CurrentExample $currentExample, WriterInterface $currentExampleWriter)
     {
         $error = array('type' => 1, 'message' => 'Hello');
         $currentExample->getCurrentExample()->willReturn('Hello');
