+++ src/PhpSpec/Formatter/WriterInterface.php
@@ -1,0 +1,21 @@
+<?php
+
+/*
+ * This file is part of PhpSpec, A php toolset to drive emergent
+ * design by specification.
+ *
+ * (c) Marcello Duarte <marcello.duarte@gmail.com>
+ * (c) Konstantin Kudryashov <ever.zet@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace PhpSpec\Formatter;
+
+use PhpSpec\Message\CurrentExample;
+
+interface WriterInterface
+{
+    public function displayFatal(CurrentExample $currentExample, $error = null);
+}
